Working jsTree FileTree
The FileTree is working, it works by giving it a directory
and then you can only explore below that directory.
the tree fetches the next ones down when you query
(open a node) for them. so it doesn't have to do
an exhaustive search, which is much faster.
It also has some nice options like checkboxes, showing
or not showing icons, dots, etc.

also, you can choose to do one more multiple selections.
Also figured out how to make the Tree not take up more
space than its own div (and eat into other panel widgets
or plots). It does this by setting the height of the
container the jstree div is in, and then setting
some styles on the jstree div, namely setting
overflow: scroll is important.

I cannot for the life of me get it to compile as typescript
. I have tried so hard making a seperate model typescript file
and running panel build, and no matter what I do, I can't get it
to work. I add all of the different properties to the python
model (__javescript__, __js_require__, __js_skip__, and in
different combinations). and I add it to an index.ts and export
them. bokeh says it builds it. But the browser console says
it can't find the model. So I don't know. Right now I am
using __implementation__ = TypeScript(str_of_my_ts_code)