Bug fixes for directly setting value
the typescript sets the model.data property with attributes that it
shouldn't have if you don't specify in the get_json function.
So you can't just do model.data = get_json() and call it good.
You have to specify args to get_json that remove those attributes, such as no_li_attr, and others.

Also a funny hacky way to get around you can't always directly
deepycopy a bokey property (and not knowing any method to just get the values)
is to just iteracte through them and extract them out that way.
such as [l for l in self.data]
or even self.data[:]

that can be deepcopied then since it is now just a list and not
some weird bokeh property