Major improvements and bugfixes
Generalizes to the FileTree to move core logic to a base Tree class to make it easier to subclass.
The functions that add children nodes on open and do a recursive search when a deep value is given
not in the tree cause of lazy loading (but a great great grandchild of something loaded)
are moved to the base class, and the only thing needed to use these are supplying callables that
supply children or parents. In the case of the `FileTree` class, this is done for you, and is the
normal directory searching stuff. But if you wanted to make a sql Tree for example, you
could write 1 or 2 optional callables that do DB queries. Or you could make a Tree that
pulls from a NetCDF, or generates new nodes programmatically. An example is supplied that
creates nodes randomly.

You can also now just use the Tree class and supply full upfront static data that doesn't
lazy load with no optional callables. This will work too.

A bug was also fixed where you directly supplied a much outside of the lazy loaded tree,
and it wasn't properly selecting it, and deselecting old nodes. There is still currently
a bug where it opens all neighboring folders. But that should be an easy fix.